third_party/WebKit/LayoutTests/http/tests/inspector/network/load-resource-for-frontend-expected.txt


Running: testLocalResource
Loading resource from file:///
Failed: Can not load local file

Running: testNetworkResource
Loading resource from http://localhost:8080/inspector/network/resources/resource.php
Success: Hello world
